refactor: Rename SessionManager to App, moved back into src/
