+++ src/components/mod.rs
@@ -1,9 +1,7 @@
-pub mod session_manager;
-
-pub use super::components::session_manager::*;
+pub mod window;
+pub use super::components::window::Window;
 
 mod popups;
 mod views;
-mod window;
 
-use super::components::{popups::*, views::*, window::*};
+use super::components::{popups::*, views::*};
